# campusmarket/backend/services/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CampusTrade API
  version: 1.0.0
  description: Production API contract for the CampusTrade campus marketplace.
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /api/listings:
    get:
      tags: [Listings]
      summary: Browse active listings
      parameters:
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: campus_zone, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        '200': { description: Listings, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Listing' } } } } }
    post:
      tags: [Listings]
      summary: Publish a listing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ListingInput' } } }
      responses:
        '201': { description: Listing created, content: { application/json: { schema: { $ref: '#/components/schemas/Listing' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/search:
    get:
      tags: [Search]
      summary: Full-text and faceted marketplace search
      parameters:
        - { in: query, name: q, schema: { type: string, maxLength: 100 } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: min_price, schema: { type: integer, minimum: 0 } }
        - { in: query, name: max_price, schema: { type: integer, minimum: 0 } }
        - { in: query, name: sort, schema: { type: string, enum: [relevance, newest, price_asc, price_desc, views] } }
      responses:
        '200': { description: Search results }
  /api/offers:
    post:
      tags: [Commerce]
      summary: Submit an offer on a listing
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [listingId, amount], properties: { listingId: { type: string, format: uuid }, amount: { type: integer, minimum: 1 } } } } }
      responses:
        '201': { description: Offer created }
  /api/conversations:
    get:
      tags: [Messaging]
      summary: List the current user's conversations
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Conversations }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/notifications:
    get:
      tags: [Notifications]
      summary: List the current user's notifications
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Notifications }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/ai/price-suggestion:
    post:
      tags: [Commerce]
      summary: Ask Claude to explain a deterministic marketplace price estimate
      description: |
        Uses completed CampusTrade sales to calculate the numeric estimate and confidence band.
        When at least three comparable sales exist, Claude gives one qualitative sentence about why
        the fixed range fits, using only normalized aggregate facts and never repeating a number.
        Claude never sets the price.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, condition]
              properties:
                category:
                  type: string
                  enum: [Accessories, Books, Bracelets, Clothing, Cosmetics, Electronics, Fruit Salad, Housing, Juice, Liquid Soap, Other, Pancake/Cake, Perfumes, Services, Shawarma, Shoes, Textbooks]
                condition: { type: string, minLength: 1, maxLength: 40 }
      responses:
        '200':
          description: Deterministic estimate with Claude explanation, or an honest low-data result without a provider claim
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [suggestion, range, confidence, count, explanation, provider, model]
                    properties:
                      suggestion: { type: integer, minimum: 1, description: Deterministic estimate in FCFA }
                      range:
                        type: object
                        required: [min, max]
                        properties:
                          min: { type: integer, minimum: 1 }
                          max: { type: integer, minimum: 1 }
                      confidence: { type: string, const: high }
                      count: { type: integer, minimum: 3 }
                      explanation: { type: string, minLength: 10, maxLength: 320 }
                      provider: { type: string, const: anthropic }
                      model: { type: string }
                  - type: object
                    required: [suggestion, confidence, count, message]
                    properties:
                      suggestion: { type: 'null' }
                      confidence: { type: string, const: low }
                      count: { type: integer, minimum: 0, maximum: 2 }
                      message: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429':
          description: The authenticated user has exhausted the shared Claude quota or the higher-volume gateway abuse guard
          content:
            application/json:
              schema:
                type: object
                required: [error, code]
                properties:
                  error: { type: string }
                  code: { type: string, enum: [CLAUDE_RATE_LIMITED, AI_GUIDANCE_ABUSE_LIMITED] }
        '503':
          description: Price aggregates, shared quota storage, or Claude are temporarily unavailable
          content:
            application/json:
              schema:
                type: object
                required: [error, code]
                properties:
                  error: { type: string }
                  code: { type: string, enum: [PRICE_DATA_UNAVAILABLE, CLAUDE_UNAVAILABLE] }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: string } } }
...
    ListingInput:
      type: object
      required: [title, price_fcfa, category, condition, campus_zone]
      properties:
        title: { type: string, minLength: 3, maxLength: 120 }
        description: { type: string, maxLength: 4000 }
        price_fcfa: { type: integer, minimum: 0 }
        category: { type: string }
        condition: { type: string }
        campus_zone: { type: string }
        images: { type: array, maxItems: 8, items: { type: string, format: uri } }
    Listing:
      allOf:
        - { $ref: '#/components/schemas/ListingInput' }
        - type: object
          required: [id, status]
          properties:
            id: { type: string, format: uuid }
            seller_id: { type: string, format: uuid }
            status: { type: string, enum: [active, sold, removed, expired] }